New Topic: Etale Cohomology/Homotopy!
added some more sections that i plan to fill out, hope there is time.
added section on etale morphisms.
section on sheaves will change, will focus on sheaves on the etale site.
added section on simplicial sets, with the end goal of defining the
etale homotopy type. this means I need to write enought category theory
so that the phrase "pro-object of the homotopy category of simplicial
sets" makes sense to the reader.`